Moving automation peers to separate files
